readings/promptflow/deployment/requirements.txt
azure-ai-ml==1.22.0
azure-identity==1.19.0
azureml-mlflow==1.58.0.post3
